pyproject.toml
[project]
name = "apisecurityengine"
version = "0.1.0"
description = "API Security Engine - automated API security testing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=13.7.0",
]

[project.scripts]
ase = "apisecurityengine.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "PIE", "PT", "SIM", "PTH", "RUF"]

[tool.mypy]
strict = true
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
